test(drive-tasks): consolidate tests into single file to fix mock collision
